Fixed issue with RTCP jitter calculation on some 64 bit platforms/compilers, thanks Alexander Sbitnev
git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/trunk@28314 023b2edf-31b2-4de3-b41e-bca80c47788f

+++ src/rtp/rtp_session.cxx
@@ -790,7 +790,7 @@ OpalRTPSession::SendReceiveStatus OpalRTPSession::OnReceiveData(RTP_DataFrame & 
 
         // As per RFC3550 Appendix 8
         diff *= GetJitterTimeUnits(); // Convert to timestamp units
-        long variance = diff - lastTransitTime;
+        long variance = (long)diff - (long)lastTransitTime;
         lastTransitTime = diff;
         if (variance < 0)
           variance = -variance;
